+++ ui-react/packages/atlasmap-ui/src/mapper/index.ts
@@ -3,4 +3,3 @@ export * from './MapperContextToolbar';
 export * from './MapperControlBar';
 export * from './MapperViewToolbar';
 export * from './MappingDetails';
-export * from './MapperUtilsToolbar';
